fix(strategist): mkdir before save, float challenge level, accurate comments
Address final review: create the checkpoint dir before model.save (a
short run may never hit the first save_freq checkpoint), annotate
_challenge_level as float since it feeds win_prob's float math, and
correct the win-prob comments in the forced-outcome tests.

+++ env/strategist_env.py
@@ -49,10 +49,10 @@ class StrategistEnv(gym.Env):
         self.challenge_idx = 0
         self.steps = 0
 
-    def _challenge_level(self) -> int:
+    def _challenge_level(self) -> float:
         # Clamp: after the last win challenge_idx == len, but _obs still runs.
         idx = min(self.challenge_idx, len(CHALLENGE_LEVELS) - 1)
-        return CHALLENGE_LEVELS[idx]
+        return float(CHALLENGE_LEVELS[idx])
 
     def _obs(self) -> np.ndarray:
         cl = self._challenge_level()
